Fixed narrowing cast in fonts.h

+++ src/rendering/fonts.h
@@ -87,7 +87,7 @@ namespace GLPL {
 				// Store characters
 				Character character = {texture,glm::ivec2(face->glyph->bitmap.width, face->glyph->bitmap.rows),
 										glm::ivec2(face->glyph->bitmap_left, face->glyph->bitmap_top),
-										face->glyph->advance.x};
+										static_cast<unsigned int>(face->glyph->advance.x)};
 				this->Characters.insert(std::pair<GLchar, Character>(c,character));
 			}
 			glBindTexture(GL_TEXTURE_2D, 0);
